autopiloot/requirements.txt
yt-dlp>=2024.7.1
google-api-python-client>=2.141.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.0
google-cloud-firestore>=2.16.0
slack_sdk>=3.31.0
openai>=1.43.0
assemblyai>=0.33.0
langfuse>=2.36.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
tiktoken>=0.11.0
# Firebase Functions dependencies
functions-framework>=3.5.0
firebase-admin>=6.5.0
firebase-functions>=0.1.0

# RAG and Memory
zep-python>=2.0.0

# Agency Swarm Framework
agency-swarm>=0.7.2

# Additional dependencies for tools
pathlib>=1.0.1
typing-extensions>=4.0.0

# Drive Agent text extraction dependencies
PyPDF2>=3.0.1
python-docx>=0.8.11
pandas>=2.0.3

# Strategy Agent NLP dependencies
scikit-learn>=1.3.0
nltk>=3.8.1
pytz>=2025.2
numpy>=2.3.3
